+++ CHANGELOG.md
@@ -9,4 +9,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Agarwal's greedy algorithm (`curvesimp.agarwal`).
+- Agarwal's greedy algorithm (`curvesimplify.agarwal`).